requirements.txt
openai>=1.0.0
streamlit>=1.27.0
faiss-cpu
python-dotenv
tqdm
fpdf
tiktoken

